fix: the location of CustomAuthenticationEntryPoint and CustomAccessDeniedHandler has been refactored
